python/requirements.txt
fastapi
uvicorn
pyngrok
pillow
python-docx
transformers
torch
python-multipart
pytesseract
pdf2image
whisper
openai-whisper
gTTS
pydub
sacremoses
googletrans==4.0.0-rc1
soundfile
librosa
deepfilternet
